# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.BookTicketRequest:
    properties:
      promo_code:
        description: opsional
        type: string
      schedule_id:
        type: string
      seat_id:
        items:
          type: string
        type: array
    required:
    - schedule_id
    - seat_id
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@cinemticket.com
    name: Cinema Ticketing Support
  description: REST API untuk sistem manajemen tiket bioskop. Mendukung manajemen
    studio, film, jadwal tayang, kursi, booking tiket, transaksi, promo diskon, dan
    laporan penjualan.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Cinema Ticketing API
  version: "1.0"
paths:
...
  /report/daily:
    get:
      description: Admin dapat melihat laporan penjualan tiket pada tanggal tertentu,
        termasuk total tiket, total pendapatan, dan breakdown per film dan studio.
      parameters:
      - description: 'Tanggal laporan (format: YYYY-MM-DD). Default: hari ini'
        example: "2026-05-20"
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Laporan penjualan harian
      tags:
      - Report
  /report/monthly:
    get:
      description: Admin dapat melihat laporan penjualan tiket selama satu bulan penuh,
        termasuk breakdown harian, per film, dan per studio.
      parameters:
      - description: 'Tahun laporan (format: YYYY). Default: tahun ini'
        example: "2026"
        in: query
        name: year
        type: string
      - description: 'Bulan laporan (1-12). Default: bulan ini'
        example: "5"
        in: query
        name: month
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Laporan penjualan bulanan
      tags:
      - Report
  /ticket:
    post:
      consumes:
      - application/json
      description: Memesan satu atau lebih kursi untuk jadwal film tertentu. Bisa
        menyertakan kode promo opsional untuk mendapatkan diskon. Tiket dibuat dengan
        status 'pending' dan akan otomatis dibatalkan jika tidak dibayar dalam 15
        menit.
      parameters:
      - description: Detail booking tiket
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.BookTicketRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Booking tiket
      tags:
      - Ticket
  /ticket/available-seats/{schedule_id}:
    get:
      description: Menampilkan daftar kursi yang belum dipesan untuk jadwal film tertentu.
      parameters:
      - description: Schedule ID (UUID)
        in: path
        name: schedule_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Kursi tersedia
      tags:
      - Ticket
  /ticket/history:
    get:
      description: Menampilkan semua riwayat pembelian tiket milik user yang sedang
        login.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Riwayat transaksi user
      tags:
      - Ticket
  /transaction:
    get:
      description: Mengembalikan seluruh data transaksi. Hanya bisa diakses oleh admin.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Lihat semua transaksi
      tags:
      - Transaction
  /transaction/{transaction_id}:
    get:
      description: Mengembalikan detail satu transaksi berdasarkan ID-nya.
      parameters:
      - description: Transaction ID (UUID)
        in: path
        name: transaction_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Detail transaksi
      tags:
      - Transaction
  /transaction/{transaction_id}/cancel:
    post:
      description: Membatalkan transaksi yang masih berstatus pending. Semua tiket
        terkait juga ikut dibatalkan dan email notifikasi dikirim ke user.
      parameters:
      - description: Transaction ID (UUID)
        in: path
        name: transaction_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Batalkan transaksi
      tags:
      - Transaction
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"